Allow for use of cURL if there is no wget.
Added the ability to use cURL if wget is not found on the box. Saw #5 and agree that that was a poor implementation but this allows for handling of either wget or cURL. 

If the options for cURL aren't the ones that would give the desired output feel free to change them.